Parent a method (JAL-code) overridden by the child method (author).
Pure cars use gas. Pure electric cars do not use gas.
So two attributes were added to the car class:
gas_refill = a warning distance in mileage for refilling the gas task
max_range = the maximum incremental range for incremental milage
updates.

When using update_odometer, the method will compare the mileage.
If the mileage is between the gas_refill and max_range a message
is printing out warning the user to refill the gas tank. See
method fill_gas_tank(self).

For the electric car, a duplicate fill_gas_tank(self) method is created.
This method is overrides the parent method when a refill is required.

The parent code example is not provided by the author of the book.
Hence the parent method is created by JAL-code.