Added pi_string.py: paired with pi_million_digits.txt
Reads pi with 1000002 characters. 2 are "3."
1000000 are the digits after the period.

The code joins the lines of the txt file and prints out the
first 52 characters in the string.

Also there is a input to find if your birthday is pi.

Added a script which increments pi from 2 to 53 characters
using two lists and a dictionary.